requirements.txt
# Core Web Framework
fastapi==0.115.12
uvicorn==0.34.2
python-multipart==0.0.20
python-dotenv==1.1.0
pydantic==2.11.4
pydantic-settings==2.11.0

# LLM & AI
openai==1.109.1
langchain==0.2.15
langchain-community==0.2.15
langchain-core<0.3.0
langchain-openai==0.1.25
langchain-experimental<0.3.0
tavily-python==0.7.12
tiktoken==0.7.0

# Embeddings & RAG
sentence-transformers>=3.3.1
huggingface-hub>=0.23.0

# Vector Stores
pinecone>=5.0.0
chromadb==1.2.1

# Database
pymongo==4.6.1
motor>=3.4.0

# Authentication & Security
firebase-admin==7.1.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# HTTP & Utilities
requests==2.32.3
httpx>=0.24.0

# Caching & Progress
cachetools>=5.3.0
tqdm>=4.65.0

# Document Processing
pypdf==4.3.1
unstructured==0.15.9
unstructured-client==0.25.5
networkx==3.3
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth>=2.0.0
python-jose[cryptography]==3.3.0

# Encryption & Security
cryptography==41.0.7

# Blockchain
psycopg2-binary==2.9.9
sarvamai
pydub

sarvamai

# Agent Lightning (Training & Optimization)

# Fixed Dependencies
numpy<2